Restructured folder structure
* All the widgets that belong to a specific view are now inside the folder with the view's name.
