+++ lib/views/filter_view.dart
@@ -3,11 +3,11 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:space_x_tracker/widgets/filters/filter_types.dart';
 import 'package:space_x_tracker/providers/launch_provider.dart';
-import 'package:space_x_tracker/widgets/filters/filter_grid_view.dart';
-import 'package:space_x_tracker/widgets/filters/filter_sub_title.dart';
+import 'package:space_x_tracker/widgets/filter/filter_grid_view.dart';
+import 'package:space_x_tracker/widgets/filter/filter_sub_title.dart';
 import 'package:space_x_tracker/widgets/flash_message.dart';
+import 'package:space_x_tracker/widgets/filter/filter_types.dart';
 
 class FilterView extends StatefulWidget {
   const FilterView({Key? key}) : super(key: key);
