+++ lib/widgets/home/launch_card.dart
@@ -3,11 +3,11 @@ import 'package:intl/intl.dart';
 import 'package:space_x_tracker/custom_color_scheme.dart';
 import 'package:space_x_tracker/providers/models/launch.dart';
 import 'package:space_x_tracker/providers/models/launch_details_arguments.dart';
-import 'package:space_x_tracker/views/launch_details_view.dart';
-import 'package:space_x_tracker/widgets/icon_row.dart';
-import 'package:space_x_tracker/widgets/patch.dart';
-import 'package:space_x_tracker/widgets/status_text.dart';
-import 'package:space_x_tracker/widgets/widget_row.dart';
+import 'package:space_x_tracker/views/rocket_details_view.dart';
+import 'package:space_x_tracker/widgets/home/icon_row.dart';
+import 'package:space_x_tracker/widgets/home/patch.dart';
+import 'package:space_x_tracker/widgets/home/status_text.dart';
+import 'package:space_x_tracker/widgets/home/widget_row.dart';
 
 class LaunchCard extends StatelessWidget {
   final Launch launch;
@@ -55,7 +55,7 @@ class LaunchCard extends StatelessWidget {
     };
 
     return InkWell(
-      onTap: () => Navigator.of(context).pushNamed(LaunchDetailsView.routeName, arguments: LaunchDetailsArguments(launch.rocket!)),
+      onTap: () => Navigator.of(context).pushNamed(RocketDetailsView.routeName, arguments: LaunchDetailsArguments(launch.rocket!)),
       child: Card(
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(20.0),
