+++ lib/widgets/home/sorting/sorting_bottom_sheet.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:space_x_tracker/widgets/flash_message.dart';
-import 'package:space_x_tracker/widgets/sorting/sorting_bottom_sheet_item.dart';
+import 'package:space_x_tracker/widgets/home/sorting/sorting_bottom_sheet_item.dart';
 
-import '../../providers/launch_provider.dart';
+import '../../../providers/launch_provider.dart';
 
 class SortingBottomSheet extends StatefulWidget {
   const SortingBottomSheet({Key? key}) : super(key: key);
