+++ test/widget/home_view_test.dart
@@ -12,7 +12,7 @@ import 'package:space_x_tracker/project_theme.dart';
 import 'package:space_x_tracker/providers/launch_provider.dart';
 import 'package:space_x_tracker/views/filter_view.dart';
 import 'package:space_x_tracker/views/home_view.dart';
-import 'package:space_x_tracker/widgets/launch_card.dart';
+import 'package:space_x_tracker/widgets/home/launch_card.dart';
 
 import 'home_view_test.mocks.dart';
 import 'launch_data.dart';
